% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{data}
\name{datos}
\alias{datos}
\title{datos dataset}
\format{An object of class "data.frame".}
\source{
IDEAM 2015 Database.
}
\usage{
data(datos)
}
\description{
datos is a dataset of daily rainfall data taken in some of the
Colombian Climate Stations for the period of 2010.01.01 to 2010.01.31
The Weather Stations are located in the coffe region mostly.
}
\examples{
data(datos)
}
\keyword{dataset}
