% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_tmin.R
\docType{data}
\name{datos_tmin}
\alias{datos_tmin}
\title{datos_tmin dataset}
\format{An object of class "data.frame".}
\source{
IDEAM 2015 Database.
}
\usage{
data(datos_tmin)
}
\description{
datos_tmin is a dataset of daily minimum temperature data taken in some of the
Colombian Climate Stations for the period of 1980.01.01 to 1980.01.31.
The Weather Stations are located in the coffe region mostly.
}
\examples{
data(datos_tmin)
}
\keyword{dataset}
