% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegRAIN.R
\name{RegRAIN}
\alias{RegRAIN}
\title{RegRAIN Spatial Interpolator for regionalized daily rainfall and air temperature surfaces.}
\usage{
RegRAIN(datos, dem, ini, fin, crossv)
}
\arguments{
\item{datos}{Object of class "data.frame" that contains the input rainfall data including the columns:
climate station code,date, longitude, latitude and rainfall data.}

\item{dem}{The Digital Elevation Model (DEM) that will be used in the interpolation.}

\item{ini}{The initial day of interpolation.}

\item{fin}{The final day of interpolation.}

\item{crossv}{When "TRUE" a cross validation process will be performed and additional to the RegRAIN raster a plot showing "goodness of fit" indicators will be exported to the working directory. If "FALSE" only the RegRAIN Interpolation raster will be generated.}
}
\description{
This function apply the method RegRAIN to interpolate daily rainfall and air temperature data from climate stations.
}
\details{
For defining the arguments, the following should be considered:


datos       is the object of class "data.frame" containing the daily rainfall data and must
            include the following columns:

CODIGO:     With the station code.

FECHA:      with the date of the data in the format YEAR-MONTH-DAY.
            If this column is not of class date and in the format YEAR-MONTH-DAY, the interpolation could be wrong.

LONGITUD:   with the longitudes or X coordinates of the climate station.

LATITUD:    with the latitudes or Y coordinates of the climate station.

PPT:        with the daily rainfall data in mm (use "." for dec). NA data should be set as "99999" or as "NA" for a proper recognition.

Run  View(datos) to see the example data.

dem:         is a raster class object. It should contain the projection information and we recommend to use tif format. The extent
            of the raster will be the same as the RegRAIN interpolation result extent.

ini:         is the initial day of interpolation. Note that the first day is the same as the first date you have in your data object. You should not use dates here, just the number that corresponds
            with the initial date you want to interpolate.

fin:         is the final day of interpolation. Note that the last day is the same as the final date you have in your data object. You should not use dates here, just the number that corresponds
            with the final date you want to interpolate.

crossv:      is of class "logical". If set "TRUE" a cross validation process for each RegRAIN raster will be performed. This process will take some aditional time depending on the number of
            climate stations.
}
\section{About RegRAIN}{


RegRAIN is a regionalized rain model, that integrates the use of digital elevation models,
their transformations (slope and aspect) and homogenized climate information recorded on surface climate stations.
The calculation is based on a Multiple Linear Regression (RLM).

                        yi = a1*xi1 +a2*xi2 + a3*xi3 + a4*xi4 +a5*xi5 + Ei 	Equation 1

   Where,

   yi = weather variable in the station "i".

   aik = Regression coefficient in the climate station "i" for the weather variable "k".

   xik = k variables in the climate station "i" with xi1 = 1.

   Ei  = residue.


 The 5(k) variables that correspond to geographic and terrain factors are

 - Latitude and longitude (decimal degrees or meters).

 - Elevation (obtained from DEM in meters).

 - Slope and Aspect for each climate station (obtained from DEM in degrees).
}

\section{References}{

Alzate-Velásquez, D.F., Araujo-Carrilo, G.A., Rojas-Barbosa, E.O., Gómez-Latorre, D.A., & Martínez-Maldonado, F.E. (2018).
Interpolacion Regnie para lluvia y temperatura en las regiones Andina, Caribe y Pacífica de Colombia. 
Colombia Forestal, 21(1), 102-118.
            
Abteilung Hydrometeorologie. (2013). REGNIE (Regionalisierte Niederschlage): Verfahrensbeschreibung & Nutzeranleitung.
Offenbach: Deutscher Wetterdienst - DWD. 9 p.

Rauthe, M., Steiner, H., U. Riediger, A., Mazurkiewicz, A., & Gratzki, A. (2013). A Central European precipitacion
climatology - Part I: Generation and validation of a high-resolution gridded daily data set (HYRAS). Meteorologische
Zeitschrift, 22(3), 235-256.

Nunez Lopez, D., Trevino Garza, E., Reyes Gomez, V., Munoz Robles, C., Aguirre Calderon, O., & Jimenez Perez, J. (2014).  
Uso de modelos de regresion para interpolar espacialmente la precipitacion media mensual en la cuenca del rio Conchos. 
Revista mexicana de ciencias agricolas, 5 (2), 201-213.
}

\examples{
#load the digital elevation model to use in the interpolation
dem <- raster(system.file("extdata", "demwgs84.tif", package="RegRAIN"))
#load the daily rainfall data that will be interpolated
datos
#Run the RegRAIN function. Here the rainfall data between the
#first (ini) and the 5th (fin) date will be interpolated, using
#the dem pixel resolution (0.0091 decimal degrees).
RegRAIN(datos, dem, 1, 2, crossv = FALSE)

}
\author{
Diego Alzate.
E-mail: diegoalzatev@gmail.com
}
